fix(okr): zero-target KRs report correct progress and completion
KeyResult::progress() returned 0.0 for any KR with target_value == 0,
and is_complete() returned true for any current_value >= 0. For
"minimize" KRs like "No critical errors" (target 0, unit count) this
meant progress was stuck at 0% forever while completion was always
true — even with errors present.

Now for zero-target KRs: current <= 0 → progress 1.0 and complete;
current > 0 → progress 0.0 and not complete. KRs with target > 0 are
unchanged. Tests extracted to tests.rs/tests_model.rs to satisfy the
50-line file budget, with two new regression tests for both branches.

CodeTether-Provenance-ID: ctprov_e62dade875406d26661dc6fe13f85030
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: 0e65c04e-72b1-4cfa-b596-afaaff2c7155
